Make launching programs from menu work
Adds python derivations for python-qubesdb and python-qubesagent. There is
probably a way to make the python packages as part of the main derivation,
but I don't know how.

If the user has set DPI-related settings through environment.variables, they
will not affect programs launched through the menu. Launching another terminal
from the first one is a workaround for now.